requirements.txt
docker-py==0.4.0
nose==1.3.3